Add new file for testing functions
