+++ src/booking-utilities.js
@@ -1,0 +1,17 @@
+
+
+const getTotalCost = (rooms) => {
+	const totalCost = rooms.reduce((cost, currRoom) => {
+		cost += currRoom.costPerNight;
+		return cost
+	}, 0)
+	const formattedCost = totalCost.toLocaleString('en-US', {
+		style: 'currency',
+		currency: 'USD'
+	});
+	return formattedCost
+}
+
+export {
+    getTotalCost,
+}
