+++ test/booking-utilities-test.js
@@ -1,0 +1,26 @@
+import chai from 'chai';
+
+import { sampleCustomers } from '../src/data/sample-customers';
+import { sampleRooms } from '../src/data/sample-rooms';
+import { sampleBookings } from '../src/data/sample-bookings';
+import { getTotalCost } from '../src/booking-utilities';
+
+
+const expect = chai.expect;
+
+
+
+// TESTS
+
+describe('Retrieve cost', function() {
+  // let customer1;
+
+  // beforeEach(() => {
+    
+  // })
+  it('should return a cost given an array of room objects', function() {
+    // const totalCost = getTotalCost(sampleRooms)
+    // console.log(totalCost)
+    // expect(totalCost).to.be.a.string;
+  });
+});
