--- test/scripts-test.js
@@ -1,24 +1,0 @@
-import chai from 'chai';
-import { sampleCustomers } from '../src/data/sample-customers';
-import { sampleRooms } from '../src/data/sample-rooms';
-import { sampleBookings } from '../src/data/sample-bookings';
-
-const expect = chai.expect;
-
-// 
-
-
-
-// TESTS
-
-describe('Get Customer', function() {
-  let customer1;
-
-  beforeEach(() => {
-    
-  })
-  it('should return a customer', function() {
-    
-    expect(customer1.name).to.equal("Reki Cyan");
-  });
-});
